% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steve.R
\name{dms2dec}
\alias{dms2dec}
\title{Convert degrees minutes seconds string to decimal degrees}
\usage{
dms2dec(s)
}
\arguments{
\item{s}{string of 1-3 coordinate chunks separated by a non-text elements.}
}
\value{
float
}
\description{
Will automatically look for W, S, N, E to identify positive or negative orientation
}
\examples{
string <- "113°30′04.24 W"
dms2dec(string)

}
