% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steve.R
\name{fromclip}
\alias{fromclip}
\title{copied excel to data.frame}
\usage{
fromclip(...)
}
\value{
data.frame of tab-separated array copied from clipboard
}
\description{
Helper function for transfering between excel and R
}
