% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steve.R
\name{toclip}
\alias{toclip}
\title{Write table to clipboard}
\usage{
toclip(d)
}
\arguments{
\item{d}{any object writable to write.table}
}
\value{
tab separated table copied to clipboard
}
\description{
Helper function for transfering between excel and R
}
